reorganize code, note forward curr not done yet.
